008-slack-approval-process/IAM-permission-request/requirements.txt
goblet-gcp==0.5.0
slack_sdk==3.13.0
google-cloud-secret-manager==2.4.0